requirements.txt
# HTTP API 통신 (DB증권 REST API Oauth2 토큰 발급 및 매매 주문 통신용)
requests>=2.31.0

# 데이터 조작 및 분석 (trade_history.csv, setting.csv 등 파일 입출력 및 전략 계산)
pandas>=2.0.0
numpy>=1.24.0

# 환경 변수 및 설정 관리 (config.py에서 AppKey, AppSecretKey 등 민감한 정보의 안전한 로드)
python-dotenv>=1.0.0

# 유닛 테스트 프레임워크 (tests/ 폴더의 test_generate_buy_orders.py 등 테스트 코드 실행용)
pytest>=7.4.0